+++ asf_search/search/search.py
@@ -1,3 +1,4 @@
+import time
 from typing import Union, Sequence, Tuple
 from copy import copy
 import datetime
@@ -12,6 +13,8 @@ def search(
         beamMode: Union[str, Sequence[str]] = None,
         beamSwath: Union[str, Sequence[str]] = None,
         campaign: Union[str, Sequence[str]] = None,
+        circle: Tuple[float, float, float] = None,
+        linestring: Sequence[float] = None,
         maxDoppler: float = None,
         minDoppler: float = None,
         end: Union[datetime.datetime, str] = None,
@@ -55,6 +58,7 @@ def search(
     :param beamMode: The beam mode used to acquire the data.
     :param beamSwath: Encompasses a look angle and beam mode.
     :param campaign: For UAVSAR and AIRSAR data collections only. Search by general location, site description, or data grouping as supplied by flight agency or project.
+    :param circle: Search by circle defined by list of three floats: [longitude, latitude, radius in meters]
     :param maxDoppler: Doppler provides an indication of how much the look direction deviates from the ideal perpendicular flight direction acquisition.
     :param minDoppler: Doppler provides an indication of how much the look direction deviates from the ideal perpendicular flight direction acquisition.
     :param end: End date of data acquisition. Supports timestamps as well as natural language such as "3 weeks ago"
@@ -94,10 +98,15 @@ def search(
     results = ASFSearchResults([])
     
     # The last page will be marked as complete if results sucessful
+    perf = time.time()
     for page in search_generator(opts=opts):
+        ASF_LOGGER.warning(f"Page Time Elapsed {time.time() - perf}")
         results.extend(page)
         results.searchComplete = page.searchComplete
         results.searchOptions = page.searchOptions
+        perf = time.time()
+    
+    results.raise_if_incomplete()
     
     try:
         results.sort(key=lambda p: p.get_sort_keys(), reverse=True)